revert move of block manager
