+++ packages/edit-post/src/plugins/index.js
@@ -11,6 +11,7 @@ import { addQueryArgs } from '@wordpress/url';
  * Internal dependencies
  */
 import CopyContentMenuItem from './copy-content-menu-item';
+import ManageBlocksMenuItem from './manage-blocks-menu-item';
 import KeyboardShortcutsHelpMenuItem from './keyboard-shortcuts-help-menu-item';
 import ToolsMoreMenuGroup from '../components/header/tools-more-menu-group';
 import WelcomeGuideMenuItem from './welcome-guide-menu-item';
@@ -22,6 +23,7 @@ registerPlugin( 'edit-post', {
 				<ToolsMoreMenuGroup>
 					{ ( { onClose } ) => (
 						<>
+							<ManageBlocksMenuItem onSelect={ onClose } />
 							<MenuItem
 								role="menuitem"
 								href={ addQueryArgs( 'edit.php', {
