+++ packages/edit-post/src/plugins/manage-blocks-menu-item/index.js
@@ -1,0 +1,34 @@
+/**
+ * WordPress dependencies
+ */
+import { MenuItem } from '@wordpress/components';
+import { withDispatch } from '@wordpress/data';
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+/**
+ * Internal dependencies
+ */
+import { store as editPostStore } from '../../store';
+
+export function ManageBlocksMenuItem( { openModal } ) {
+	return (
+		<MenuItem
+			onClick={ () => {
+				openModal( 'edit-post/manage-blocks' );
+			} }
+		>
+			{ __( 'Block Manager' ) }
+		</MenuItem>
+	);
+}
+
+export default withDispatch( ( dispatch ) => {
+	const { openModal } = dispatch( editPostStore );
+
+	return {
+		openModal,
+	};
+} )( ManageBlocksMenuItem );
